Adding FiltersTypes component
